+++ ChangeLog.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - New `@compile_mode("unsupported")` for modules that do not support TorchScript
 - flake8 settings have been added to `setup.cfg` for improved code style
 - `TensorProduct.visualize()` can now plot weights
+- `basis='bessel'` option to `math.soft_one_hot_linspace`
 
 ### Changed
 - `o3.TensorProduct` now uses `torch.fx` to generate it's code
@@ -27,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `normalize2mom` is now a `torch.nn.Module`
 - rename arguments `set_ir_...` into `filter_ir_...`
 - Renamed `e3nn.nn.Gate` argument `irreps_nonscalars` to `irreps_gated`
+- Renamed `e3nn.o3.TensorProduct` arguments `x1, x2` to `x, y`
 
 ### Fixed
 - `nn.Gate` was crashing when the number of scalars or gates was zero